[#57]refactor: Nregister form을 Regsiter로 옮김
