+++ client/src/App.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
 
+import Register from './pages/register/Register';
+
 function App() {
-  return <div className="App"></div>;
+  return <div className="App"><Register/></div>; // merge할 때 register 제거
 }
 
 export default App;
