+++ client/src/index.tsx
@@ -4,7 +4,7 @@ import App from "./App";
 import { GlobalStyle } from "./styles/GlobalStyles";
 import { ThemeProvider } from "styled-components";
 import { theme } from "./styles/Colors";
-import "@fortawesome/fontawesome-free/js/all.js";
+// import "@fortawesome/fontawesome-free/js/all.js"; // merge할 때 주석 해제
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
